feat: decompose App.jsx - extract AppHeader, AppSidebar, DetailsPanel, custom hooks and models
