+++ VAH.Frontend/src/App.jsx
@@ -7,9 +7,10 @@ import useAssets from './hooks/useAssets';
 import useTags from './hooks/useTags';
 import useSignalR from './hooks/useSignalR';
 import useUndoRedo from './hooks/useUndoRedo';
-import { staticUrl } from './api/client';
-import * as smartCollectionsApi from './api/smartCollectionsApi';
-import CollectionTree from './components/CollectionTree';
+import useSmartCollections from './hooks/useSmartCollections';
+import AppHeader from './components/AppHeader';
+import AppSidebar from './components/AppSidebar';
+import DetailsPanel from './components/DetailsPanel';
 import AssetDisplayer from './components/AssetDisplayer';
 import UploadArea from './components/UploadArea';
 import CollectionBrowser from './components/CollectionBrowser';
@@ -85,26 +86,16 @@ function AppLayout() {
   const { tags, createTag, deleteTag, getAssetTags, setAssetTags } = useTags();
 
   // ------- smart collections -------
-  const [smartCollections, setSmartCollections] = useState([]);
-  const [activeSmartCollection, setActiveSmartCollection] = useState(null);
-  const [smartItems, setSmartItems] = useState([]);
+  const {
+    smartCollections,
+    activeSmartCollection,
+    smartItems,
+    handleSelectSmartCollection,
+    clearSmartCollection,
+  } = useSmartCollections(isAuthenticated);
 
-  const fetchSmartCollections = useCallback(async () => {
-    try {
-      const defs = await smartCollectionsApi.fetchSmartCollections();
-      setSmartCollections(defs);
-    } catch (e) { console.error('Smart collections error:', e); }
-  }, []);
-
-  useEffect(() => { if (isAuthenticated) fetchSmartCollections(); }, [isAuthenticated, fetchSmartCollections]);
-
-  const handleSelectSmartCollection = useCallback(async (sc) => {
-    setActiveSmartCollection(sc);
-    try {
-      const result = await smartCollectionsApi.fetchSmartCollectionItems(sc.id);
-      setSmartItems(result.items || []);
-    } catch (e) { console.error('Smart collection items error:', e); }
-  }, []);
+  // ------- smart collection item fetch -------
+  // (handled inside useSmartCollections hook)
 
   // ------- real-time sync -------
   const signalRHandlers = useMemo(() => ({
@@ -147,17 +138,6 @@ function AppLayout() {
   const showFolderActions = selectedCollection && (selectedCollection.type === 'image' || selectedCollection.type === 'default');
   const showLinkAction = selectedCollection && (selectedCollection.type === 'link' || selectedCollection.type === 'default');
 
-  const formatDate = (dateStr) => {
-    if (!dateStr) return '—';
-    const d = new Date(dateStr);
-    return d.toLocaleDateString('vi-VN', { day: '2-digit', month: '2-digit', year: 'numeric', hour: '2-digit', minute: '2-digit' });
-  };
-
-  const getContentTypeLabel = (type) => {
-    const map = { image: 'Hình ảnh', link: 'Liên kết', color: 'Màu sắc', folder: 'Thư mục', file: 'Tệp tin', default: 'Tệp tin' };
-    return map[type] || type;
-  };
-
   const handleSelectCollection = (collection, path = []) => {
     selectCollection(collection, path);
     setSelectedAssetId(null);
@@ -170,99 +150,41 @@ function AppLayout() {
     setSelectedAssetId(null);
   };
 
+  const handleAddTag = async (assetId) => {
+    const tagName = prompt('Nhập tên tag:');
+    if (!tagName) return;
+    try {
+      const tag = await createTag(tagName);
+      await setAssetTags(assetId, [tag.id]);
+      refreshItems();
+    } catch (e) {
+      console.error('Tag error:', e);
+    }
+  };
+
   return (
     <div className="app">
       {/* ====== TOP HEADER ====== */}
-      <header className="app-header">
-        <div className="header-left">
-          <button className="hamburger-btn" title="Menu">
-            <svg viewBox="0 0 24 24" width="20" height="20" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round"><line x1="3" y1="6" x2="21" y2="6"/><line x1="3" y1="12" x2="21" y2="12"/><line x1="3" y1="18" x2="21" y2="18"/></svg>
-          </button>
-          <div className="logo">
-            <div className="logo-icon">
-              <svg viewBox="0 0 24 24" width="20" height="20" fill="none" stroke="white" strokeWidth="2.2" strokeLinecap="round" strokeLinejoin="round"><circle cx="12" cy="12" r="10"/><path d="M12 8v8M8 12h8"/></svg>
-            </div>
-            <span className="logo-text">Visual Asset Hub</span>
-          </div>
-        </div>
-
-        <div className="header-center">
-          <div className="search-wrapper">
-            <span className="search-icon-el">
-              <svg viewBox="0 0 24 24" width="16" height="16" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round"><circle cx="11" cy="11" r="8"/><line x1="21" y1="21" x2="16.65" y2="16.65"/></svg>
-            </span>
-            <input
-              type="text"
-              id="global-search"
-              name="search"
-              placeholder="Tìm kiếm..."
-              value={searchTerm}
-              onChange={(e) => handleSearchChange(e.target.value)}
-              className="search-input"
-              autoComplete="off"
-            />
-          </div>
-        </div>
-
-        <div className="header-right">
-          <button className="header-icon-btn" title="Thư mục">
-            <svg viewBox="0 0 24 24" width="20" height="20" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"><path d="M22 19a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h5l2 3h9a2 2 0 0 1 2 2z"/></svg>
-          </button>
-          <button className="header-icon-btn" title="Cài đặt">
-            <svg viewBox="0 0 24 24" width="20" height="20" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"><circle cx="12" cy="12" r="3"/><path d="M19.4 15a1.65 1.65 0 0 0 .33 1.82l.06.06a2 2 0 0 1-2.83 2.83l-.06-.06a1.65 1.65 0 0 0-1.82-.33 1.65 1.65 0 0 0-1 1.51V21a2 2 0 0 1-4 0v-.09A1.65 1.65 0 0 0 9 19.4a1.65 1.65 0 0 0-1.82.33l-.06.06a2 2 0 0 1-2.83-2.83l.06-.06A1.65 1.65 0 0 0 4.68 15a1.65 1.65 0 0 0-1.51-1H3a2 2 0 0 1 0-4h.09A1.65 1.65 0 0 0 4.6 9a1.65 1.65 0 0 0-.33-1.82l-.06-.06a2 2 0 0 1 2.83-2.83l.06.06A1.65 1.65 0 0 0 9 4.68a1.65 1.65 0 0 0 1-1.51V3a2 2 0 0 1 4 0v.09a1.65 1.65 0 0 0 1 1.51 1.65 1.65 0 0 0 1.82-.33l.06-.06a2 2 0 0 1 2.83 2.83l-.06.06A1.65 1.65 0 0 0 19.4 9a1.65 1.65 0 0 0 1.51 1H21a2 2 0 0 1 0 4h-.09a1.65 1.65 0 0 0-1.51 1z"/></svg>
-          </button>
-          <button className="header-icon-btn" title="Thông báo">
-            <svg viewBox="0 0 24 24" width="20" height="20" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"><path d="M18 8A6 6 0 0 0 6 8c0 7-3 9-3 9h18s-3-2-3-9"/><path d="M13.73 21a2 2 0 0 1-3.46 0"/></svg>
-            <span className="notif-dot"></span>
-          </button>
-          <button className="header-icon-btn" title="Đăng xuất" onClick={logout}>
-            <svg viewBox="0 0 24 24" width="20" height="20" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"><path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4"/><polyline points="16 17 21 12 16 7"/><line x1="21" y1="12" x2="9" y2="12"/></svg>
-          </button>
-          <div className="user-avatar" title={user?.displayName || 'User'}>
-            {user?.displayName ? user.displayName.slice(0, 2).toUpperCase() : 'U'}
-          </div>
-        </div>
-      </header>
+      <AppHeader
+        user={user}
+        logout={logout}
+        searchTerm={searchTerm}
+        onSearchChange={handleSearchChange}
+      />
 
       {/* ====== APP BODY (Sidebar + Main + Details) ====== */}
       <div className="app-body">
         {/* ====== LEFT SIDEBAR ====== */}
-        <aside className="app-sidebar">
-          <div className="sidebar-header">
-            <h3>Tài liệu của tôi</h3>
-            <button className="add-collection-btn" onClick={() => {
-              const name = prompt('Tên collection:');
-              if (name) handleCreateCollection(name);
-            }}>+</button>
-          </div>
-          <div className="sidebar-scroll">
-            <CollectionTree 
-              collections={collections}
-              selectedCollection={selectedCollection}
-              onSelectCollection={(c) => handleSelectCollection(c, [c])}
-              onCreateCollection={handleCreateCollection}
-              onDeleteCollection={handleDeleteCollection}
-            />
-
-            {/* Smart Collections */}
-            {smartCollections.length > 0 && (
-              <div className="smart-collections-section">
-                <h4 className="sidebar-section-title">Bộ sưu tập thông minh</h4>
-                {smartCollections.map((sc) => (
-                  <button
-                    key={sc.id}
-                    className={`smart-collection-item ${activeSmartCollection?.id === sc.id ? 'active' : ''}`}
-                    onClick={() => { handleSelectCollection(null, []); handleSelectSmartCollection(sc); }}
-                  >
-                    <span className="sc-icon">{sc.icon}</span>
-                    <span className="sc-name">{sc.name}</span>
-                    <span className="sc-count">{sc.count}</span>
-                  </button>
-                ))}
-              </div>
-            )}
-          </div>
-        </aside>
+        <AppSidebar
+          collections={collections}
+          selectedCollection={selectedCollection}
+          smartCollections={smartCollections}
+          activeSmartCollection={activeSmartCollection}
+          onSelectCollection={handleSelectCollection}
+          onCreateCollection={handleCreateCollection}
+          onDeleteCollection={handleDeleteCollection}
+          onSelectSmartCollection={handleSelectSmartCollection}
+        />
 
         {/* ====== MAIN CONTENT AREA ====== */}
         <main className="app-main">
@@ -436,7 +358,7 @@ function AppLayout() {
             <div className="smart-collection-view">
               <div className="main-toolbar">
                 <div className="toolbar-breadcrumbs">
-                  <button className="breadcrumb-item" onClick={() => { setActiveSmartCollection(null); setSmartItems([]); }}>
+                  <button className="breadcrumb-item" onClick={clearSmartCollection}>
                     <svg viewBox="0 0 24 24" width="14" height="14" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" style={{marginRight: 4, verticalAlign: 'middle'}}><path d="M3 9l9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z"/><polyline points="9 22 9 12 15 12 15 22"/></svg>
                     Trang chủ
                   </button>
@@ -490,100 +412,12 @@ function AppLayout() {
 
         {/* ====== RIGHT DETAILS PANEL ====== */}
         {selectedAsset && (
-          <aside className="details-panel">
-            <div className="details-panel-header">
-              <h3>Tài liệu</h3>
-              <button className="details-close-btn" onClick={() => setSelectedAssetId(null)} title="Đóng">
-                <svg viewBox="0 0 24 24" width="18" height="18" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round"><line x1="18" y1="6" x2="6" y2="18"/><line x1="6" y1="6" x2="18" y2="18"/></svg>
-              </button>
-            </div>
-
-            {/* Large icon / thumbnail */}
-            <div className="details-preview">
-              {selectedAsset.contentType === 'image' ? (
-                <img
-                  src={staticUrl(selectedAsset.filePath)}
-                  alt={selectedAsset.fileName}
-                  onError={(e) => { e.target.style.display = 'none'; }}
-                />
-              ) : selectedAsset.contentType === 'color' ? (
-                <div style={{ width: 120, height: 120, borderRadius: 12, backgroundColor: selectedAsset.filePath, boxShadow: '0 4px 16px rgba(0,0,0,0.3)' }} />
-              ) : (
-                <div className="details-preview-icon">
-                  {selectedAsset.isFolder ? '📁' : selectedAsset.contentType === 'link' ? '🔗' : '📄'}
-                </div>
-              )}
-            </div>
-
-            {/* Info section */}
-            <div className="details-info">
-              <h3 className="details-info-title">{selectedAsset.fileName}</h3>
-              
-              <div className="details-info-section">
-                <h4>Thông tin</h4>
-                <div className="details-info-row">
-                  <span className="details-info-label">Title</span>
-                  <span className="details-info-value">{selectedAsset.fileName}</span>
-                </div>
-                <div className="details-info-row">
-                  <span className="details-info-label">Đơn giữ/đặn</span>
-                  <span className="details-info-value">{selectedCollection?.name || '—'}</span>
-                </div>
-                <div className="details-info-row">
-                  <span className="details-info-label">Tải lên</span>
-                  <span className="details-info-value">{formatDate(selectedAsset.createdAt)}</span>
-                </div>
-                <div className="details-info-row">
-                  <span className="details-info-label">Loại tệp</span>
-                  <span className="details-info-value">{getContentTypeLabel(selectedAsset.contentType)}</span>
-                </div>
-                <div className="details-info-row">
-                  <span className="details-info-label">Thư mục gốc</span>
-                  <span className="details-info-value" style={{ display: 'flex', alignItems: 'center', gap: 4 }}>
-                    {selectedAsset.isFolder ? '—' : '■'}
-                  </span>
-                </div>
-
-                {/* Tags section */}
-                <div className="details-tags-section">
-                  <h4>Tags</h4>
-                  <div className="details-tags-list">
-                    {selectedAsset.tags && selectedAsset.tags.split(',').filter(Boolean).map((tag, i) => (
-                      <span key={i} className="tag-badge">{tag.trim()}</span>
-                    ))}
-                    {(!selectedAsset.tags || selectedAsset.tags.split(',').filter(Boolean).length === 0) && (
-                      <span className="details-info-value" style={{ opacity: 0.5 }}>Chưa có tag</span>
-                    )}
-                  </div>
-                  <button className="btn-small" onClick={async () => {
-                    const tagName = prompt('Nhập tên tag:');
-                    if (!tagName) return;
-                    try {
-                      const tag = await createTag(tagName);
-                      await setAssetTags(selectedAsset.id, [tag.id]);
-                      refreshItems();
-                    } catch (e) {
-                      console.error('Tag error:', e);
-                    }
-                  }}>+ Tag</button>
-                </div>
-              </div>
-            </div>
-
-            {/* Preview section for images/videos */}
-            {selectedAsset.contentType === 'image' && (
-              <div className="details-preview-section">
-                <h4>Chỉ xem</h4>
-                <div className="details-full-preview">
-                  <img
-                    src={staticUrl(selectedAsset.filePath)}
-                    alt={selectedAsset.fileName}
-                    onError={(e) => { e.target.parentElement.style.display = 'none'; }}
-                  />
-                </div>
-              </div>
-            )}
-          </aside>
+          <DetailsPanel
+            asset={selectedAsset}
+            collectionName={selectedCollection?.name}
+            onClose={() => setSelectedAssetId(null)}
+            onAddTag={handleAddTag}
+          />
         )}
       </div>
